backend/pyproject.toml
[project]
name = "nyayalens-backend"
version = "0.1.0"
description = "NyayaLens Backend ML Pipeline"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.4.22",
    "fastapi>=0.110.0",
    "google-cloud-translate>=3.15.3",
    "google-cloud-vision>=3.7.2",
    "groq>=0.5.0",
    "langchain-community>=0.0.33",
    "langchain-core>=0.1.0",
    "langchain-groq>=0.1.0",
    "langchain>=0.1.16",
    "numpy>=1.26.0",
    "pdfplumber>=0.11.0",
    "pydantic>=2.7.0",
    "pymongo>=4.6.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.9",
    "scikit-learn>=1.4.0",
    "sentence-transformers>=2.7.0",
    "spacy>=3.7.4",
    "transformers>=4.40.0",
    "uvicorn[standard]>=0.29.0",
]

[tool.uv]
package = false
